feat: loongcollector support windows build (#2082)
* feat: windows build

* feat: add windows workflows

* fix: windows build script

---------

+++ core/common/MachineInfoUtil.h
@@ -58,13 +58,13 @@ struct ECSMeta {
 
 private:
     std::array<char, ID_MAX_LENGTH> mInstanceID{};
-    size_t mInstanceIDLen = 0UL;
+    size_t mInstanceIDLen = (size_t)0;
 
     std::array<char, ID_MAX_LENGTH> mUserID{};
-    size_t mUserIDLen = 0UL;
+    size_t mUserIDLen = (size_t)0;
 
     std::array<char, ID_MAX_LENGTH> mRegionID{};
-    size_t mRegionIDLen = 0UL;
+    size_t mRegionIDLen = (size_t)0;
 
     friend class InstanceIdentityUnittest;
 };
@@ -104,7 +104,7 @@ struct Hostid {
 
 private:
     std::array<char, ID_MAX_LENGTH> mId{};
-    size_t mIdLen = 0UL;
+    size_t mIdLen = (size_t)0;
 
     Type mType;
