+++ core/models/MetricValue.cpp
@@ -81,7 +81,7 @@ size_t DataSize(const MetricValue& value) {
         [](auto&& arg) {
             using T = decay_t<decltype(arg)>;
             if constexpr (is_same_v<T, monostate>) {
-                return 0UL;
+                return (size_t)0;
             } else {
                 return arg.DataSize();
             }
