+++ core/protobuf/sls/LogGroupSerializer.cpp
@@ -38,13 +38,13 @@ const string METRIC_LABELS_KEY_VALUE_SEPARATOR = "#$#";
  *      Number of bytes required.
  */
 static inline size_t uint32_size(uint32_t v) {
-    if (v < (1UL << 7)) {
+    if (v < ((size_t)1 << 7)) {
         return 1;
-    } else if (v < (1UL << 14)) {
+    } else if (v < ((size_t)1 << 14)) {
         return 2;
-    } else if (v < (1UL << 21)) {
+    } else if (v < ((size_t)1 << 21)) {
         return 3;
-    } else if (v < (1UL << 28)) {
+    } else if (v < ((size_t)1 << 28)) {
         return 4;
     } else {
         return 5;
@@ -110,7 +110,7 @@ void LogGroupSerializer::StartToAddLog(size_t size) {
 
 void LogGroupSerializer::AddLogTime(uint32_t logTime) {
     // limit logTime's min value, ensure varint size is 5, which is 1978-07-05 05:24:16
-    static uint32_t minLogTime = 1UL << 28;
+    static uint32_t minLogTime = (uint32_t)1 << 28;
     if (logTime < minLogTime) {
         logTime = minLogTime;
     }
